+++ CHANGELOG.md
@@ -9,10 +9,15 @@
   buffer.
   ([#6](https://github.com/asomers/divbuf/pull/6))
 
+- `impl TryFrom<DivBufShared> for Vec<u8>` to extract the backing Vec from a
+  `DivBufShared` if there are no other DivBufs for the same `DivBufShared`.
+  ([#17](https://github.com/asomers/divbuf/pull/17))
+
 ### Changed
-- MSRV has been raised to 1.31.0
+- MSRV has been raised to 1.40.0
   ([#8](https://github.com/asomers/divbuf/pull/8))
   ([#10](https://github.com/asomers/divbuf/pull/10))
+  ([#17](https://github.com/asomers/divbuf/pull/17))
 
 ### Fixed
 - Eliminated usage of `compare_and_swap`, deprecated in Rust 1.50.0.